inventory_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
asyncpg
pydantic
python-dotenv
psycopg2-binary